Refactored more
